* 'develop' of codeup.aliyun.com:5eda397fd1d1abe63b55f034/mall-to/laravel-admin-enhance:
  调整
  多语言支持
  fix bug
  升级调整

+++ src/Form/Field/QiniuFile.php
@@ -38,7 +38,10 @@ class QiniuFile extends Field\File
             $this->name = $filePath;
         }
 
-        $this->destroy();
+        if(is_string($this->original))
+        {
+            $this->destroy();
+        }
 
         return $filePath;
     }
